+++ tests/views/test_projects_view.py
@@ -2,20 +2,23 @@
 from briefy.leica import models
 from conftest import BaseTestView
 
+import pytest
 
+
+@pytest.mark.usefixtures('create_dependencies')
 class TestProjectView(BaseTestView):
-    """Test ProjecService view."""
+    """Test ProjectService view."""
 
     base_path = '/projects'
+    dependencies = [
+        (models.Customer, 'data/customers.json')
+    ]
     file_path = 'data/projects.json'
     model = models.Project
     UPDATE_SUCCESS_MESSAGE = ''
     NOT_FOUND_MESSAGE = ''
     update_map = {
-        'briefing': 'The briefy',
-        'currency_set_price': 'USD',
-        'abstract': 'The abstract',
-        'company_id': '66e07fc8-f237-4bab-a883-14234b258513',
-        'name': 'Other Name',
-        'set_price': 200,
+        'description': 'The briefy',
+        'customer_id': 'c2034c1b-0a40-4b84-9ace-54b958f64ed4',
+        'title': 'Other Name'
     }