+++ HISTORY.rst
@@ -2,7 +2,12 @@
 History
 =======
 
-0.1.0 (Unreleased)
----------------------------
+1.0.0 (Unreleased)
+------------------
 
-* TODO
+* LEICA-24: Clean up Job and Project models. (rudaporto)
+* Add Metadata and Briefy Roles mixins to Job and Project. (rudaporto)
+* LEICA-23: Add new Customer model and link to Project. (rudaporto)
+* Update all postman tests and add into the project. (rudaporto)
+* Recreate initial alembic migrations. (rudaporto)
+* Update all tests and test data to fit the changes in the models. (rudaporto)